Allow float value instead of int
There are parts of the application that would send float values to
this method.

This change intends to address a PHP Deprecation in PHP 8.1.

+++ lib/Color/Gradient.php
@@ -79,7 +79,7 @@ use function array_fill;
         return new static(...$colors);
     }
 
-    public function colorAtPercentile(int $percentile): Color
+    public function colorAtPercentile(float $percentile): Color
     {
         $percentile = $percentile < 0 ? 0 : min(100, abs($percentile));
